backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
elasticsearch==8.11.0
openai==1.99.1
aiohttp==3.12.15
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-instrumentation-openai-v2==2.1b0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-instrumentation-elasticsearch==0.57b0
opentelemetry-instrumentation-httpx==0.57b0
opentelemetry-instrumentation-azure==0.57b0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-util-http==0.57b0
APScheduler==3.10.4
pydantic_settings==2.10.1